+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.0] - 2022-09-02
+
 ### Fixed
 
 - Can run the same [un]stow[-root] flag multiple times
@@ -68,7 +70,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - New Github actions to auto package and release to Github and PyPI
 - New Github action to update CHANGELOG after release
 
-[unreleased]: https://github.com/ghassan0/stowd/compare/v0.1.0...HEAD
+[unreleased]: https://github.com/ghassan0/stowd/compare/v0.2.0...HEAD
+[0.2.0]: https://github.com/ghassan0/stowd/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/ghassan0/stowd/compare/v0.0.2...v0.1.0
 [0.0.2]: https://github.com/ghassan0/stowd/compare/v0.0.1...v0.0.2
 [0.0.1]: https://github.com/ghassan0/stowd/releases/tag/v0.0.1